[ansible] Adds roles for wordpress, mysql and git
Adds new roles for wordpress and mysql in docker. Also a role for
git to clone down the wordpress repo you would like to have.
Removes the site.yml in favor of a file for every role.
Because I apply the correct role with my own script this makes
more sense.